divide project on a modules
