+++ index.html
@@ -15,6 +15,6 @@
         </div>
         <div id="hint">blablabla</div>
     </div>
-    <script src="scripts/main.js"></script>
+    <script src="scripts/main.js" type="module"></script>
 </body>
 </html>
