+++ scripts/utils/generateRandomNumber.js
@@ -1,0 +1,3 @@
+const generateRandomNumber = (size, gap) => Math.floor(Math.random() * (size - (gap * 2))) + gap;
+
+export {generateRandomNumber};
