+++ scripts/utils/generateTreasureCoords.js
@@ -1,0 +1,8 @@
+import {generateRandomNumber} from './generateRandomNumber.js';
+
+const generateTreasureCoords = ({width, height}, gap) => ({
+    x: generateRandomNumber(width, gap),
+    y: generateRandomNumber(height,gap)
+});
+
+export {generateTreasureCoords};
